Rename tcn_parse.go to tcnparse.go
